Refactor simple distributed caching support []
The simple distributed caching in Bazel used ConcurrentMap as a blob store. It is incorrect to use an overloaded interface for this purpose.

This change defines a SimpleBlobStore interface that only has put(), get() and containsKey()
methods and allows a simple implementation of a blob store as remote cache for Bazel.

Also updated documentation to summarize the options available in the remote spwan strategy.

There is no functional change.

TESTED=shell integration tests

PiperOrigin-RevId: 151439467